+++ HISTORY.rst
@@ -11,6 +11,7 @@ PlatformIO 4.0
 
 * Print `debug tool <http://docs.platformio.org/en/latest/plus/debugging.html#tools-debug-probes>`__ name for the active debugging session
 * Do not shutdown PIO Home Server for "upgrade" operations (`issue #2784 <https://github.com/platformio/platformio-core/issues/2784>`_)
+* Improved computing of project check sum (structure, configuration) and avoid unnecessary rebuilding
 * Fixed an issue with incorrect escaping of Windows slashes when using `PIO Unified Debugger <http://docs.platformio.org/en/latest/plus/debugging.html>`__ and "piped" openOCD
 
 4.0.0 (2019-07-10)